requirement.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
beautifulsoup4==4.12.2
bleach==6.0.0
Brotli==1.0.9
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
cryptography==41.0.2
cssselect2==0.7.0
Django==4.2.3
django-ckeditor==6.7.0
django-downloadview==2.3.0
django-js-asset==2.1.0
filelock==3.12.2
fire==0.5.0
fonttools==4.41.1
html5lib==1.1
idna==3.4
imageio==2.31.1
Jinja2==3.1.2
lazy_loader==0.3
lxml==4.9.3
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
ninja==1.11.1
numpy==1.25.1
opencv-python==4.8.0.74
opencv-python-headless==4.8.0.74
oscrypto==1.3.0
packaging==23.1
pdf2docx==0.5.6
pdfkit==1.0.0
PDFNetPython3==9.5.0.post2
Pillow==10.0.0
pyclipper==1.3.0.post4
pycparser==2.21
pydyf==0.7.0
pyHanko==0.20.0
pyhanko-certvalidator==0.23.0
PyMuPDF==1.22.5
pypdf==3.14.0
PyPDF2==3.0.1
pyphen==0.14.0
pypng==0.20220715.0
python-bidi==0.4.2
python-docx==0.8.11
PyWavelets==1.4.1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==3.6.13
requests==2.31.0
scikit-image==0.21.0
scipy==1.11.1
shapely==2.0.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
svglib==1.5.1
sympy==1.12
termcolor==2.3.0
tifffile==2023.7.18
tinycss2==1.2.1
torchvision==0.15.2
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
uritools==4.0.1
urllib3==2.0.4
webencodings==0.5.1
xhtml2pdf==0.2.11
zopfli==0.2.2
